Updated to latest SystemsRx and changed optimistic sub handling
The original fix worked ok but this new approach will remove the excess memory overhead and cleanup without allocation if needed which hopefully will reduce memory usage and bring back more speed.

This has also allowed certain code to be streamlined in the setup phase too which makes it easier to maintain.